agents turn towards player and advances

+++ Rogue-Robots/DOGEngine/src/ECS/Component.h
@@ -45,7 +45,7 @@ namespace DOG
 
 	struct NetworkPlayerComponent
 	{
-		int playerId;
+		i8 playerId;
 	};
 
 	struct NetworkTransform
